requirements.txt
fastapi==0.110.0
matplotlib==3.8.3
numpy==1.26.4
opencv_python_headless==4.9.0.80
pandas==2.2.1
Pillow==10.2.0
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1
uvicorn==0.27.1